+++ slo_generator/backends/datadog.py
@@ -21,6 +21,8 @@ import pprint
 
 import datadog
 
+from slo_generator import utils
+
 LOGGER = logging.getLogger(__name__)
 logging.getLogger("datadog.api").setLevel(logging.ERROR)
 
@@ -61,35 +63,48 @@ class DatadogBackend:
         start = timestamp - window
         end = timestamp
         query_good = measurement["query_good"]
-        query_valid = measurement["query_valid"]
+
+        if measurement.get("query_bad"):
+            query = measurement.get("query_bad")
+        elif measurement.get("query_valid"):
+            query = measurement.get("query_valid")
+        else:
+            raise ValueError("One of `query_bad` or `query_valid` is required.")
+
         query_good = self._fmt_query(
             query_good,
             window,
             operator,
             operator_suffix,
         )
-        query_valid = self._fmt_query(
-            query_valid,
-            window,
-            operator,
-            operator_suffix,
-        )
+
         good_event_query = self.client.Metric.query(
             start=start,
             end=end,
             query=query_good,
         )
-        valid_event_query = self.client.Metric.query(
+
+        query = self._fmt_query(
+            query,
+            window,
+            operator,
+            operator_suffix,
+        )
+
+        event_query = self.client.Metric.query(
             start=start,
             end=end,
-            query=query_valid,
+            query=query,
         )
-        LOGGER.debug(f"Result good: {pprint.pformat(good_event_query)}")
-        LOGGER.debug(f"Result valid: {pprint.pformat(valid_event_query)}")
+
         good_event_count = DatadogBackend.count(good_event_query)
-        valid_event_count = DatadogBackend.count(valid_event_query)
-        bad_event_count = valid_event_count - good_event_count
-        return (good_event_count, bad_event_count)
+        event_count = DatadogBackend.count(event_query)
+        if measurement.get("query_valid"):
+            event_count = event_count - good_event_count
+
+        LOGGER.debug(f"Good events: {good_event_count} | " f"Bad events: {event_count}")
+
+        return good_event_count, event_count
 
     def query_sli(self, timestamp, window, slo_config):
         """Query SLI value directly.
@@ -124,8 +139,9 @@ class DatadogBackend:
         """
         slo_id = slo_config["spec"]["service_level_indicator"]["slo_id"]
         from_ts = timestamp - window
-        slo_data = self.client.ServiceLevelObjective.get(id=slo_id)
-        LOGGER.debug(f"SLO data: {slo_id} | Result: {pprint.pformat(slo_data)}")
+        if utils.is_debug_enabled():
+            slo_data = self.client.ServiceLevelObjective.get(id=slo_id)
+            LOGGER.debug(f"SLO data: {slo_id} | Result: {pprint.pformat(slo_data)}")
         data = self.client.ServiceLevelObjective.history(
             id=slo_id,
             from_ts=from_ts,
@@ -137,7 +153,7 @@ class DatadogBackend:
             valid_event_count = data["data"]["series"]["denominator"]["sum"]
             bad_event_count = valid_event_count - good_event_count
             return (good_event_count, bad_event_count)
-        except (KeyError) as exception:  # monitor-based SLI
+        except KeyError as exception:  # monitor-based SLI
             sli_value = data["data"]["overall"]["sli_value"] / 100
             LOGGER.debug(exception)
             return sli_value