+++ slo_generator/migrations/migrator.py
@@ -249,6 +249,7 @@ def exporters_v1tov2(
     exp_keys = []
     for exp_path in exporters_paths:
         with open(exp_path, encoding="utf-8") as conf:
+            # pylint: disable=E1111
             content = yaml.load(conf, Loader=yaml.SafeLoader)
         exporters = content
 
@@ -284,7 +285,9 @@ def ebp_v1tov2(ebp_paths: list, shared_config: dict = {}, quiet: bool = False) -
     ebp_keys = []
     for ebp_path in ebp_paths:
         with open(ebp_path, encoding="utf-8") as conf:
+            # pylint: disable=E1111
             error_budget_policy = yaml.load(conf, Loader=yaml.SafeLoader)
+        # pylint: disable=E1133
         for step in error_budget_policy:
             step["name"] = step.pop("error_budget_policy_step_name")
             step["burn_rate_threshold"] = step.pop("alerting_burn_rate_threshold")
@@ -416,18 +419,15 @@ def report_v2tov1(report: dict) -> dict:
     """
     mapped_report: dict = {}
     for key, value in report.items():
-
         # If a metadata label is passed, use the metadata label mapping
         if key == "metadata":
             mapped_report["metadata"] = {}
             for subkey, subvalue in value.items():
-
                 # v2 `metadata.labels` attributes map to `metadata` attributes
                 # in v1
                 if subkey == "labels":
                     labels = subvalue
                     for labelkey, labelval in labels.items():
-
                         # Top-level labels like 'service_name', 'feature_name',
                         # and 'slo_name'.
                         if labelkey in METRIC_METADATA_LABELS_TOP_COMPAT: