+++ tests/unit/test_compute.py
@@ -19,7 +19,10 @@ from datadog.api import Metric, ServiceLevelObjective
 from elasticsearch import Elasticsearch
 from google.auth._default import _CLOUD_SDK_CREDENTIALS_WARNING
 from mock import MagicMock, patch
+from opensearchpy import OpenSearch
 from prometheus_http_client import Prometheus
+from splunklib import client as Splunk
+from splunklib.client import Jobs
 
 from slo_generator.backends.dynatrace import DynatraceClient
 from slo_generator.compute import compute, export
@@ -38,8 +41,10 @@ from .test_stubs import (
     mock_dt,
     mock_dt_errors,
     mock_es,
+    mock_os_client,
     mock_prom,
     mock_sd,
+    mock_splunk_oneshot,
     mock_ssm_client,
 )
 
@@ -53,6 +58,8 @@ SLO_CONFIGS_PROM = load_slo_samples("prometheus", CTX)
 SLO_CONFIGS_ES = load_slo_samples("elasticsearch", CTX)
 SLO_CONFIGS_DD = load_slo_samples("datadog", CTX)
 SLO_CONFIGS_DT = load_slo_samples("dynatrace", CTX)
+SLO_CONFIGS_SPLUNK = load_slo_samples("splunk", CTX)
+SLO_CONFIGS_OS = load_slo_samples("opensearch", CTX)
 SLO_REPORT = load_fixture("slo_report_v2.json")
 SLO_REPORT_V1 = load_fixture("slo_report_v1.json")
 EXPORTERS = load_fixture("exporters.yaml", CTX)
@@ -60,7 +67,7 @@ BQ_ERROR = load_fixture("bq_error.json")
 
 # Pub/Sub methods to patch
 PUBSUB_MOCKS = [
-    "google.cloud.pubsub_v1.gapic.publisher_client.PublisherClient.publish",
+    "google.cloud.pubsub_v1.PublisherClient.publish",
     "google.cloud.pubsub_v1.publisher.futures.Future.result",
     "google.api_core.grpc_helpers.create_channel",
 ]
@@ -74,6 +81,13 @@ SSM_MOCKS = [
 
 class TestCompute(unittest.TestCase):
     maxDiff = None
+
+    @patch.object(Jobs, "oneshot", side_effect=mock_splunk_oneshot)
+    @patch.object(Splunk, "connect", return_value=None)
+    def test_splunk_search(self, *mocks):
+        for config in SLO_CONFIGS_SPLUNK:
+            with self.subTest(config=config):
+                compute(config, CONFIG)
 
     @patch(
         "google.api_core.grpc_helpers.create_channel",
@@ -131,6 +145,12 @@ class TestCompute(unittest.TestCase):
     @patch.object(DynatraceClient, "request", side_effect=mock_dt)
     def test_compute_dynatrace(self, mock):
         for config in SLO_CONFIGS_DT:
+            with self.subTest(config=config):
+                compute(config, CONFIG)
+
+    @patch.object(OpenSearch, "search", mock_os_client)
+    def test_compute_opensearch(self):
+        for config in SLO_CONFIGS_OS:
             with self.subTest(config=config):
                 compute(config, CONFIG)